+++ lightx2v/models/networks/hunyuan/infer/transformer_infer.py
@@ -1,11 +1,12 @@
-from lightx2v.models.networks.aaa_transformer.transformer_infer import BaseTransformer
+from lightx2v.common.transformer_infer.transformer_infer import BaseTransformerInfer
 import torch
 from einops import rearrange
 from .utils_bf16 import apply_rotary_emb
 import numpy as np
 from lightx2v.common.offload.manager import WeightAsyncStreamManager
+from lightx2v.utils.envs import *
 
-class BaseHunyuanTransformer(BaseTransformer):
+class BaseHunyuanTransformerInfer(BaseTransformerInfer):
     # 1. 初始化
     def __init__(self, config):
         self.config = config
@@ -256,3 +257,37 @@ class BaseHunyuanTransformer(BaseTransformer):
             previous_out = block_cache[module_name][0]
             block_cache[module_name][0] = out
             block_cache[module_name][1] = (out - previous_out) / step_diff
+
+class HunyuanTransformerInfer(BaseHunyuanTransformerInfer):
+    def __init__(self, config):
+        super().__init__(config)
+
+    @torch.compile(disable=not CHECK_ENABLE_GRAPH_MODE())
+    def infer(self, weights, img, txt, vec, cu_seqlens_qkv, max_seqlen_qkv, freqs_cis, token_replace_vec=None, frist_frame_token_num=None):
+        # 1. 读取数组
+        index = self.scheduler.step_index
+        caching_records = self.scheduler.caching_records
+
+        # 2. 判断完全计算，或者使用缓存
+        if caching_records[index]:
+            return self.infer_calculating(weights, img, txt, vec, cu_seqlens_qkv, max_seqlen_qkv, freqs_cis,token_replace_vec, frist_frame_token_num)
+        else:
+            return self.infer_using_cache(weights, img, txt, vec, cu_seqlens_qkv, max_seqlen_qkv, freqs_cis,token_replace_vec, frist_frame_token_num)
+
+    def infer_calculating(self, weights, img, txt, vec, cu_seqlens_qkv, max_seqlen_qkv, freqs_cis, token_replace_vec=None, frist_frame_token_num=None):
+        txt_seq_len = txt.shape[0]
+        img_seq_len = img.shape[0]
+        for i in range(self.double_blocks_num):
+            img_out, txt_out, img_mod1_gate, img_mod2_shift, img_mod2_scale, img_mod2_gate,tr_img_mod1_gate, tr_img_mod2_shift, tr_img_mod2_scale, tr_img_mod2_gate, txt_mod1_gate, txt_mod2_shift, txt_mod2_scale, txt_mod2_gate = super().infer_double_block_1(weights.double_blocks[i], img, txt, vec, cu_seqlens_qkv, max_seqlen_qkv, freqs_cis, token_replace_vec, frist_frame_token_num)
+            img, txt, img_out, txt_out, img_mod2_gate, txt_mod2_gate = super().infer_double_block_2(weights.double_blocks[i], img, txt, vec, cu_seqlens_qkv, max_seqlen_qkv, freqs_cis, token_replace_vec, frist_frame_token_num, img_out, txt_out, img_mod1_gate, img_mod2_shift, img_mod2_scale, img_mod2_gate,tr_img_mod1_gate, tr_img_mod2_shift, tr_img_mod2_scale, tr_img_mod2_gate, txt_mod1_gate, txt_mod2_shift, txt_mod2_scale, txt_mod2_gate)
+            img, txt = super().infer_double_block_3(weights.double_blocks[i], img, txt, vec, cu_seqlens_qkv, max_seqlen_qkv, freqs_cis, token_replace_vec, frist_frame_token_num, img_out, txt_out, img_mod2_gate, txt_mod2_gate)
+
+        x = torch.cat((img, txt), 0)
+        for i in range(self.single_blocks_num):
+            out, mod_gate, tr_mod_gate = super().infer_single_block_1(weights.single_blocks[i], x, vec, txt_seq_len, cu_seqlens_qkv, max_seqlen_qkv, freqs_cis, token_replace_vec, frist_frame_token_num)
+            x = super().infer_single_block_2(weights.single_blocks[i], x, vec, txt_seq_len, cu_seqlens_qkv, max_seqlen_qkv, freqs_cis, out, mod_gate, tr_mod_gate, token_replace_vec, frist_frame_token_num)
+        img = x[:img_seq_len, ...]
+        return img, vec
+        
+    def infer_using_cache(self, weights, img, txt, vec, cu_seqlens_qkv, max_seqlen_qkv, freqs_cis, token_replace_vec=None, frist_frame_token_num=None):
+        pass
