--- lightx2v/models/networks/wan/infer/feature_caching/transformer_infer_3.py
@@ -1,138 +1,0 @@
-from ..transformer_infer import BaseWanTransformer
-import torch
-import math
-
-class WanTransformerInferTaylorCaching(BaseWanTransformer):
-    def __init__(self, config):
-        super().__init__(config)
-
-        self.status = "even"
-        self.blocks_cache_even = [{} for _ in range(self.blocks_num)]
-        self.blocks_cache_odd = [{} for _ in range(self.blocks_num)]
-        
-    def infer(self, weights, embed, grid_sizes, x, embed0, seq_lens, freqs, context):
-        if self.status == "even":
-            index = self.scheduler.step_index
-            caching_records = self.scheduler.caching_records
-
-            if caching_records[index]:
-                x = self.infer_calculating(weights, grid_sizes, x, embed0, seq_lens, freqs, context)
-            else:
-                x = self.infer_using_cache(weights, grid_sizes, x, embed0, seq_lens, freqs, context)
-            self.status = "odd"
-
-        else:
-            index = self.scheduler.step_index
-            caching_records_2 = self.scheduler.caching_records_2
-
-            if caching_records_2[index]:
-                x = self.infer_calculating(weights, grid_sizes, x, embed0, seq_lens, freqs, context)
-            else:
-                x = self.infer_using_cache(weights, grid_sizes, x, embed0, seq_lens, freqs, context)
-            self.status = "even"
-
-        return x
-
-    def infer_calculating(self, weights, grid_sizes, x, embed0, seq_lens, freqs, context):
-        for block_idx in range(self.blocks_num):
-            y_out, gate_msa, c_shift_msa, c_scale_msa, c_gate_msa = super().infer_block_1(weights.blocks[block_idx], grid_sizes, x, embed0, seq_lens, freqs, context)
-            if self.status == "even":
-                self.derivative_approximation(self.blocks_cache_even[block_idx],"self_attn_out",y_out)
-            else:
-                self.derivative_approximation(self.blocks_cache_odd[block_idx],"self_attn_out",y_out)
-
-            attn_out = super().infer_block_2(weights.blocks[block_idx], grid_sizes, x, embed0, seq_lens, freqs, context, y_out, gate_msa)
-            if self.status == "even":
-                self.derivative_approximation(self.blocks_cache_even[block_idx],"cross_attn_out",attn_out)
-            else:
-                self.derivative_approximation(self.blocks_cache_odd[block_idx],"cross_attn_out",attn_out)
-
-            y_out = super().infer_block_3(weights.blocks[block_idx], grid_sizes, x, embed0, seq_lens, freqs, context, attn_out, c_shift_msa, c_scale_msa)
-            if self.status == "even":
-                self.derivative_approximation(self.blocks_cache_even[block_idx],"ffn_out",y_out)
-            else:
-                self.derivative_approximation(self.blocks_cache_odd[block_idx],"ffn_out",y_out)
-
-            x = super().infer_block_4(weights.blocks[block_idx], grid_sizes, x, embed0, seq_lens, freqs, context, y_out, c_gate_msa)
-        return x
-
-    def infer_using_cache(self, weights, grid_sizes, x, embed0, seq_lens, freqs, context):
-        for block_idx in range(self.blocks_num):
-            x = self.infer_block(weights.blocks[block_idx], grid_sizes, x, embed0, seq_lens, freqs, context, block_idx)
-        return x
-
-    # 1. taylor using caching
-    def infer_block(self, weights, grid_sizes, x, embed0, seq_lens, freqs, context, i):
-        # 1. shift, scale, gate
-        if embed0.dim() == 3:
-            modulation = weights.modulation.tensor.unsqueeze(2)
-            embed0 = (modulation + embed0).chunk(6, dim=1)
-            _, _, gate_msa, _, _, c_gate_msa = [ei.squeeze(1) for ei in embed0]
-        elif embed0.dim() == 2:
-            _, _, gate_msa, _, _, c_gate_msa = (weights.modulation.tensor + embed0).chunk(6, dim=1)
-
-        # 2. residual and taylor
-        if self.status == "even":
-            out = self.taylor_formula(self.blocks_cache_even[i]["self_attn_out"])
-            out = out * gate_msa.squeeze(0)
-            x = x + out
-
-            out = self.taylor_formula(self.blocks_cache_even[i]["cross_attn_out"])
-            x = x + out
-
-            out = self.taylor_formula(self.blocks_cache_even[i]["ffn_out"])
-            out = out * c_gate_msa.squeeze(0)
-            x = x + out
-
-            return x
-
-        else:
-            out = self.taylor_formula(self.blocks_cache_odd[i]["self_attn_out"])
-            out = out * gate_msa.squeeze(0)
-            x = x + out
-
-            out = self.taylor_formula(self.blocks_cache_odd[i]["cross_attn_out"])
-            x = x + out
-
-            out = self.taylor_formula(self.blocks_cache_odd[i]["ffn_out"])
-            out = out * c_gate_msa.squeeze(0)
-            x = x + out
-
-            return x
-
-    # 1. when fully calcualted, stored in cache
-    def derivative_approximation(self, block_cache, module_name, out):
-        if module_name not in block_cache:
-            block_cache[module_name] = {0: out}
-        else:
-            step_diff = self.get_taylor_step_diff()
-
-            previous_out = block_cache[module_name][0]
-            block_cache[module_name][0] = out
-            block_cache[module_name][1] = (out - previous_out) / step_diff
-
-    def get_taylor_step_diff(self):
-        step_diff = 0
-        if self.status == "even":
-            current_step = self.scheduler.step_index
-            last_calc_step = current_step - 1
-            while last_calc_step >= 0 and not self.scheduler.caching_records[last_calc_step]:
-                last_calc_step -= 1
-            step_diff = current_step - last_calc_step
-        else:
-            current_step = self.scheduler.step_index
-            last_calc_step = current_step - 1
-            while last_calc_step >= 0 and not self.scheduler.caching_records_2[last_calc_step]:
-                last_calc_step -= 1
-            step_diff = current_step - last_calc_step
-        
-        return step_diff
-    
-    def taylor_formula(self, tensor_dict):
-        x = self.get_taylor_step_diff()
-
-        output = 0
-        for i in range(len(tensor_dict)):
-            output += (1 / math.factorial(i)) * tensor_dict[i] * (x**i)
-
-        return output
