+++ lightx2v/models/networks/wan/infer/transformer_infer.py
@@ -1,11 +1,11 @@
-from lightx2v.models.networks.aaa_transformer.transformer_infer import BaseTransformer
+from lightx2v.common.transformer_infer.transformer_infer import BaseTransformerInfer
 import torch
 from .utils import compute_freqs, compute_freqs_dist, apply_rotary_emb
 from lightx2v.common.offload.manager import WeightAsyncStreamManager
 from lightx2v.utils.envs import *
 
 
-class BaseWanTransformer(BaseTransformer):
+class BaseWanTransformerInfer(BaseTransformerInfer):
     # 1. 初始化
     def __init__(self, config):
         # 1.1 配置 + 任务 + 注意力
@@ -165,3 +165,31 @@ class BaseWanTransformer(BaseTransformer):
         cu_seqlens_q = torch.cat([q_lens.new_zeros([1]), q_lens]).cumsum(0, dtype=torch.int32)
         cu_seqlens_k = torch.cat([k_lens.new_zeros([1]), k_lens]).cumsum(0, dtype=torch.int32)
         return cu_seqlens_q, cu_seqlens_k
+    
+
+class WanTransformerInfer(BaseWanTransformerInfer):
+    def __init__(self, config):
+        super().__init__(config)
+
+    @torch.compile(disable=not CHECK_ENABLE_GRAPH_MODE())
+    def infer(self, weights, embed, grid_sizes, x, embed0, seq_lens, freqs, context):
+        # 1. 读取数组
+        index = self.scheduler.step_index
+        caching_records = self.scheduler.caching_records
+
+        # 2. 判断完全计算，或者使用缓存
+        if caching_records[index]:
+            return self.infer_calculating(weights, grid_sizes, x, embed0, seq_lens, freqs, context)
+        else:
+            return self.infer_using_cache(weights, grid_sizes, x, embed0, seq_lens, freqs, context)
+
+    def infer_calculating(self, weights, grid_sizes, x, embed0, seq_lens, freqs, context):
+        for block_idx in range(self.blocks_num):
+            y_out, gate_msa, c_shift_msa, c_scale_msa, c_gate_msa = super().infer_block_1(weights.blocks[block_idx], grid_sizes, x, embed0, seq_lens, freqs, context)
+            attn_out = super().infer_block_2(weights.blocks[block_idx], grid_sizes, x, embed0, seq_lens, freqs, context, y_out, gate_msa)
+            y_out = super().infer_block_3(weights.blocks[block_idx], grid_sizes, x, embed0, seq_lens, freqs, context, attn_out, c_shift_msa, c_scale_msa)
+            x = super().infer_block_4(weights.blocks[block_idx], grid_sizes, x, embed0, seq_lens, freqs, context, y_out, c_gate_msa)
+        return x
+
+    def infer_using_cache(self, weights, grid_sizes, x, embed0, seq_lens, freqs, context):
+        pass
