+++ lightx2v/models/networks/wan/model.py
@@ -10,9 +10,8 @@ from lightx2v.models.networks.wan.weights.transformer_weights import (
 )
 from lightx2v.models.networks.wan.infer.pre_infer import WanPreInfer
 from lightx2v.models.networks.wan.infer.post_infer import WanPostInfer
-from lightx2v.models.networks.wan.infer.feature_caching.transformer_infer_1 import WanTransformerInfer
-from lightx2v.models.networks.wan.infer.feature_caching.transformer_infer_2 import WanTransformerInferTeaCaching
-from lightx2v.models.networks.wan.infer.feature_caching.transformer_infer_3 import WanTransformerInferTaylorCaching
+from lightx2v.models.networks.wan.infer.transformer_infer import WanTransformerInfer
+from lightx2v.models.networks.wan.infer.feature_caching.transformer_infer import WanTransformerInferTeaCaching, WanTransformerInferTaylorCaching, WanTransformerInferAdaCaching
 from safetensors import safe_open
 import lightx2v.attentions.distributed.ulysses.wrap as ulysses_dist_wrap
 import lightx2v.attentions.distributed.ring.wrap as ring_dist_wrap
@@ -58,6 +57,8 @@ class WanModel:
             self.transformer_infer_class = WanTransformerInferTeaCaching
         elif self.config["feature_caching"] == "TaylorSeer":
             self.transformer_infer_class = WanTransformerInferTaylorCaching
+        elif self.config["feature_caching"] == "Ada":
+            self.transformer_infer_class = WanTransformerInferAdaCaching
         else:
             raise NotImplementedError(f"Unsupported feature_caching type: {self.config['feature_caching']}")
