+++ lightx2v/models/runners/wan/wan_runner.py
@@ -6,7 +6,7 @@ from PIL import Image
 from lightx2v.utils.registry_factory import RUNNER_REGISTER
 from lightx2v.models.runners.default_runner import DefaultRunner
 from lightx2v.models.schedulers.wan.scheduler import WanScheduler
-from lightx2v.models.schedulers.wan.feature_caching.scheduler import WanSchedulerTeaCaching, WanSchedulerTaylorCaching
+from lightx2v.models.schedulers.wan.feature_caching.scheduler import WanSchedulerTeaCaching, WanSchedulerTaylorCaching, WanSchedulerAdaCaching
 from lightx2v.utils.profiler import ProfilingContext
 from lightx2v.models.input_encoders.hf.t5.model import T5EncoderModel
 from lightx2v.models.input_encoders.hf.xlm_roberta.model import CLIPModel
@@ -95,6 +95,8 @@ class WanRunner(DefaultRunner):
             scheduler = WanSchedulerTeaCaching(self.config)
         elif self.config.feature_caching == "TaylorSeer":
             scheduler = WanSchedulerTaylorCaching(self.config)
+        elif self.config.feature_caching == "Ada":
+            scheduler = WanSchedulerAdaCaching(self.config)
         else:
             raise NotImplementedError(f"Unsupported feature_caching type: {self.config.feature_caching}")
         self.model.set_scheduler(scheduler)
